fix: shop module structure
change the shop module structure and add a new module shop_database that contains all the entities and repositories of shop
add a line in the README about swagger

linked #15
