+++ Server/shop/product/pom.xml
@@ -3,12 +3,11 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.2.0</version>
-        <relativePath/> <!-- lookup parent from repository -->
+        <groupId>com.cashmanager.server</groupId>
+        <artifactId>shop</artifactId>
+        <version>0.0.1-SNAPSHOT</version>
     </parent>
-    <groupId>com.cashmanager.server</groupId>
+
     <artifactId>product</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>product</name>
@@ -45,13 +44,7 @@
                 </dependency>
         <dependency>
             <groupId>com.cashmanager.server</groupId>
-            <artifactId>shop</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.cashmanager.server</groupId>
-            <artifactId>common</artifactId>
+            <artifactId>shop_database</artifactId>
             <version>0.0.1-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
