+++ Server/shop/product/src/main/java/com/cashmanager/server/product/ProductApplication.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
 @EnableJpaRepositories(basePackages = "com.cashmanager.server")
-@EntityScan("com.cashmanager.server.shop.entities")
+@EntityScan("com.cashmanager.server.shop_database.entities")
 public class ProductApplication {
 
     public static void main(String[] args) {
