+++ Server/shop/product/src/main/java/com/cashmanager/server/product/exceptions/ProductException.java
@@ -1,5 +1,8 @@
 package com.cashmanager.server.product.exceptions;
 
+/**
+ * Class that will be thrown when an error occur while communicated with the database
+ */
 public class ProductException extends Exception{
     public ProductException() {
         super();
