+++ Server/shop/shop_database/compose.yaml
@@ -1,0 +1,12 @@
+services:
+  postgres:
+    image: 'postgres:16.1-alpine'
+    volumes:
+      - 'db:/var/lib/postgres/data'
+      - '../../init-shop-db.sh:/docker-entrypoint-initdb.d/init-shop-db.sh'
+    env_file:
+      - ../../.env.local
+    ports:
+      - '5432:5432'
+volumes:
+  db:
