+++ Server/shop/shop_database/src/main/java/com/cashmanager/server/shop_database/ShopDatabaseApplication.java
@@ -1,0 +1,13 @@
+package com.cashmanager.server.shop_database;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ShopDatabaseApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ShopDatabaseApplication.class, args);
+    }
+
+}
