+++ Server/shop/shop_database/src/main/java/com/cashmanager/server/shop_database/repositories/ProductRepository.java
@@ -1,0 +1,9 @@
+package com.cashmanager.server.shop_database.repositories;
+
+import com.cashmanager.server.shop_database.entities.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface ProductRepository extends JpaRepository<Product, UUID> {
+}
