+++ Server/shop/shop_database/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+# WARNING!!! Currently this launching shop will create tables in bank database
+# postgresql database url for unit tests
+spring.datasource.url=jdbc:postgresql://localhost:5432/shop_db
+# postgresql database username
+spring.datasource.username=shop_user
+# postgresql database password
+spring.datasource.password=shop_password
+
