+++ Server/shop/shop_database/src/main/resources/db/migration/V1__create_products.sql
