+++ Server/shop/shop_database/src/main/resources/db/migration/V4__update_column_price.sql
