+++ Server/shop/shop_database/src/test/java/com/cashmanager/server/shop_database/ShopDatabaseApplicationTests.java
@@ -1,0 +1,13 @@
+package com.cashmanager.server.shop_database;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ShopDatabaseApplicationTests {
+
+//    @Test
+//    void contextLoads() {
+//    }
+
+}
