--- Server/shop/src/main/java/com/cashmanager/server/shop/ShopApplication.java
@@ -1,16 +1,0 @@
-package com.cashmanager.server.shop;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-
-@SpringBootApplication
-@EnableJpaRepositories
-public class ShopApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ShopApplication.class, args);
-    }
-
-}
