--- Server/shop/src/main/java/com/cashmanager/server/shop/entities/Order.java
@@ -1,50 +1,0 @@
-package com.cashmanager.server.shop.entities;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.proxy.HibernateProxy;
-
-import java.util.Date;
-import java.util.Objects;
-import java.util.Set;
-import java.util.UUID;
-
-@Getter
-@Setter
-@ToString
-@RequiredArgsConstructor
-@Builder
-@AllArgsConstructor
-@Entity
-@Table(name = "orders")
-public class Order {
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(name = "id" ,nullable = false)
-    // GenerationType instructs that a UUID for the entity should be generated automatically for us by the persistence provider.
-    private UUID id;
-    @Column( nullable = false)
-    private String status;
-    @Column( nullable = false)
-    private Date date;
-
-    @OneToMany(mappedBy = "order")
-    @ToString.Exclude
-    private Set<OrderedOrder> orderedOrders;
-
-    @Override
-    public final boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null) return false;
-        Class<?> oEffectiveClass = o instanceof HibernateProxy ? ((HibernateProxy) o).getHibernateLazyInitializer().getPersistentClass() : o.getClass();
-        Class<?> thisEffectiveClass = this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass() : this.getClass();
-        if (thisEffectiveClass != oEffectiveClass) return false;
-        Order order = (Order) o;
-        return getId() != null && Objects.equals(getId(), order.getId());
-    }
-
-    @Override
-    public final int hashCode() {
-        return this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass().hashCode() : getClass().hashCode();
-    }
-}
