--- Server/shop/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-# WOARNING!!! Currently this launching shop will create tables in bank database
-# postgresql database url for unit tests
-spring.datasource.url=jdbc:postgresql://localhost:5432/shop_db
-# postgresql database username
-spring.datasource.username=shop_user
-# postgresql database password
-spring.datasource.password=shop_password
-#Configuration for the eureka-server
-#spring.application.name=SHOP-SERVICE
-#eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
