--- Server/shop/src/test/java/com/cashmanager/server/shop/ShopApplicationTests.java
@@ -1,12 +1,0 @@
-package com.cashmanager.server.shop;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ShopApplicationTests {
-
-//    @Test
-//    void contextLoads() {
-//    }
-
-}
